Allow indicating pixel dimensions in encoder
Writes the dimension information before the IDAT chunks in the header of
the image. This avoids having to manually construct such a chunk which
could also violate the order restriction set by the specification. Also
adds a utility function which constructs the PixelDimensions from a
simple dpi parameter.

Addresses part of #116